requirements.txt
longport>=1.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
# vnpy>=3.6.0 - removed due to dependency issues with ta-lib
# qlib>=0.9.0 - removed due to compatibility issues
# pyfolio>=0.9.2 - removed due to compatibility issues
backtrader>=1.9.76
influxdb-client>=1.36.0
prometheus-client>=0.16.0
matplotlib>=3.7.0
scipy>=1.10.0
plotly>=5.14.0
# dolphindb>=1.30.21 - removed due to compatibility issues
PyQt5>=5.15.0
loguru>=0.7.0
pydantic>=2.0.0
tqdm>=4.65.0 